requirements.txt
asgiref==3.7.2
Django==5.0.1
django-cors-headers==4.3.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.2
psycopg2==2.9.9
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
Pillow==10.0.0
certifi==2023.11.17
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
django-rest-swagger==2.2.0
idna==3.6
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
openapi-codec==1.3.2
pytz==2023.3.post1
requests==2.31.0
simplejson==3.19.2
urllib3==2.1.0
drf-yasg==1.21.7
inflection==0.5.1
# Marketing ai
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
charset-normalizer==3.3.2
colorama==0.4.6
dataclasses-json==0.6.1
distro==1.8.0
exceptiongroup==1.1.3
frozenlist==1.4.0
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
iniconfig==2.0.0
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.332
langsmith==0.0.62
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.1
openai==0.28.1
packaging==23.2
pluggy==1.3.0
pydantic==2.4.2
pydantic_core==2.10.1
pytest==7.4.3
pytest-subtests==0.11.0
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.23
tenacity==8.2.3
tomli==2.0.1
tqdm==4.66.1
typing-inspect==0.9.0
yarl==1.9.2